refactor: deMeteorize sendOrderEmail func
