+++ imports/plugins/core/orders/server/index.js
@@ -1,5 +1,4 @@
 import { Meteor } from "meteor/meteor";
-import "./startup";
 import "./i18n";
 import methods from "./methods";
