+++ imports/plugins/core/orders/server/methods/shipmentShipped.js
@@ -3,8 +3,9 @@ import { Meteor } from "meteor/meteor";
 import { check } from "meteor/check";
 import Reaction from "/imports/plugins/core/core/server/Reaction";
 import rawCollections from "/imports/collections/rawCollections";
+import getGraphQLContextInMeteorMethod from "/imports/plugins/core/graphql/server/getGraphQLContextInMeteorMethod";
 import createNotification from "/imports/plugins/included/notifications/server/no-meteor/createNotification";
-import sendOrderEmail from "../util/sendOrderEmail";
+import sendOrderEmail from "../no-meteor/util/sendOrderEmail";
 import updateShipmentStatus from "../util/updateShipmentStatus";
 
 /**
@@ -29,7 +30,8 @@ export default function shipmentShipped(order, fulfillmentGroup) {
   });
 
   // Notify by email
-  sendOrderEmail(order, "shipped");
+  const context = Promise.await(getGraphQLContextInMeteorMethod(Reaction.getUserId()));
+  sendOrderEmail(context, order, "shipped");
 
   // Notify by in-app notification
   const { accountId } = order;
