--- imports/plugins/core/orders/server/startup.js
@@ -1,4 +1,0 @@
-import appEvents from "/imports/node-app/core/util/appEvents";
-import sendOrderEmail from "/imports/plugins/core/orders/server/util/sendOrderEmail";
-
-appEvents.on("afterOrderCreate", ({ order }) => sendOrderEmail(order));
